fix(time): remove redundant native timeval conversion
PosixSusecondsT is now i64 on supported native LP64 targets. Remove the obsolete cast in timeval-to-timespec conversion so the x86_64 CI Clippy gate passes without suppressing the lint or changing runtime semantics.

Validated with the complete make fmt check including all-features Clippy, make kernel, and git diff --check. Separate build and test jobs are blocked before guest execution by an expired dependency mirror TLS certificate; this change does not bypass certificate verification.

+++ kernel/src/time/mod.rs
@@ -187,7 +187,7 @@ impl From<PosixTimeval> for PosixTimeSpec {
     fn from(value: PosixTimeval) -> Self {
         PosixTimeSpec {
             tv_sec: value.tv_sec,
-            tv_nsec: value.tv_usec as i64 * 1000,
+            tv_nsec: value.tv_usec * 1000,
         }
     }
 }
